chapter_102/src/test/java/ru/alazarev/socket/fileManager/ServerTest.java
package ru.alazarev.socket.fileManager;

import org.junit.Ignore;
import org.junit.Test;
import ru.alazarev.socket.oracle.Server;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.File;
import java.io.IOException;
import java.net.Socket;
import java.util.List;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.*;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

public class ServerTest {

    private static final String LS = System.getProperty("line.separator");

    @Test
    public void serverTest() throws IOException {
        String input = "child";
        String extend = null;
        Socket socket = mock(Socket.class);
        ByteArrayOutputStream out = new ByteArrayOutputStream();
        when(socket.getInputStream()).thenReturn(new ByteArrayInputStream(input.getBytes()));
        when(socket.getOutputStream()).thenReturn(out);
        new Server(socket).start();
        assertThat(out.toString(), is(extend));
    }
}